add _curses_panel module
The name of the panel type was changed in 3.8
